kickstarts: feature: Add readonly root filesystem
The stateless part is used as the 'config' partition.

This commit modifies the necessary files in order to support
readonly filesystem on Redpesk ARZ.
To do so it adds some definitions of folder binding into read-write
partitions (for example SSH).

Add the "readonly-root" package which is mandatory to run the system
in readonly.

Signed-off-by: Antoine Treboutte <antoine.treboutte@iot.bzh>
Signed-off-by: Pierre Marzin <pierre.marzin@iot.bzh>