ks: partition: use auto mode and set disk msdos
Restore usage of part ks command.
Use a fictiv partition to created an offset at the beggining of
the partition scheme. This partition is then deleted during post
steps and fstab file is update consequently.
The fstab is modified too in order to use UUID in place of part
name to simplify detection during boot of the created rp-image.
The bootloader command is modified in order to match the correct
device name within the mock. This is to install the bootloader
into the correct device. As we are in a mock and /dev is shared
with host... We need to be sure to not modify a critical host
device.
Moreover, use gpt disklabel as much as possible and improve setup
of EFI partitionning for x86_64 images.

Revert "kickstart: rp-part: Add new boot partitionning step"
This reverts commit 00a561622ddacb9373fbcd46388e73f56f76b214.

Revert "kickstart: x86: setup boot like arm boards"
This reverts commit 6ec4290bf00a729abeb35e338d6b331e4ac90903.

Signed-off-by: Pierre Marzin <pierre.marzin@iot.bzh>